refactor: join node & children components
