--- src/app/resume-page/collapsible-tree/collapsible-tree-children/collapsible-tree-children.component.html
@@ -1,28 +1,0 @@
-<ul
-  *ngIf="children.length"
-  [attr.id]="id"
-  [@expanded]="!isCollapsible || isExpanded"
-  [ngClass]="
-    isCollapsible && _platformService.isServer
-      ? [DISPLAY_FLEX_IF_NO_SCRIPT_CLASS]
-      : undefined
-  "
-  [style.visibility]="
-    isCollapsible && _platformService.isBrowser ? 'hidden' : undefined
-  "
-  [style.display]="_platformService.isServer ? 'none' : undefined"
-  [attr.data-is-nested]="depth > 0"
->
-  <li *ngFor="let item of children">
-    <!--suppress AngularUndefinedTag, AngularUndefinedBinding
-    WebStorm can't detect import with `forwardRef`:
-    https://youtrack.jetbrains.com/issue/WEB-57162 -->
-    <app-collapsible-tree-node
-      [line]="item"
-      [isCollapsibleFn]="isCollapsibleFn"
-      [depth]="depth + 1"
-      [parent]="parent"
-    >
-    </app-collapsible-tree-node>
-  </li>
-</ul>
