--- src/app/resume-page/collapsible-tree/collapsible-tree-children/collapsible-tree-children.component.scss
@@ -1,18 +1,0 @@
-@use 'margins';
-
-:host {
-  ul {
-    display: flex;
-    flex-direction: column;
-    gap: margins.$m;
-    overflow: hidden;
-
-    &[data-is-nested='true'] {
-      margin-left: margins.$l;
-
-      > li:first-child {
-        margin-top: margins.$m;
-      }
-    }
-  }
-}
