+++ src/app/resume-page/collapsible-tree/collapsible-tree.component.spec.ts
@@ -1,24 +1,22 @@
 import { ComponentFixture } from '@angular/core/testing'
 
 import {
-  COLLAPSIBLE_TREE_CHILDREN_COMPONENT_FORWARD_REF,
-  CollapsibleTreeNodeComponent,
+  CollapsibleTreeComponent,
   IsCollapsibleFn,
-} from './collapsible-tree-node.component'
+} from './collapsible-tree.component'
 import {
   PLATFORM_SERVICE,
   PlatformService,
-} from '../../../common/platform.service'
+} from '../../common/platform.service'
 import { componentTestSetup } from '@test/helpers/component-test-setup'
-import { MockComponent, MockProvider, MockService } from 'ng-mocks'
+import { MockProvider } from 'ng-mocks'
 import {
   MOCK_BROWSER_PLATFORM_SERVICE,
   MOCK_SERVER_PLATFORM_SERVICE,
 } from '@test/helpers/platform-service'
 import { By } from '@angular/platform-browser'
 import { MATERIAL_SYMBOLS_SELECTOR } from '@test/helpers/material-symbols'
-import { DescriptionLine } from '../../../metadata'
-import { CollapsibleTreeChildrenComponent } from '../collapsible-tree-children/collapsible-tree-children.component'
+import { DescriptionLine } from '../../metadata'
 import { byComponent } from '@test/helpers/component-query-predicates'
 import {
   ATTRIBUTE_ARIA_CONTROLS,
@@ -26,11 +24,20 @@ import {
   ATTRIBUTE_ARIA_HIDDEN,
 } from '@test/helpers/aria'
 import { getReflectedAttribute } from '@test/helpers/get-reflected-attribute'
-import { expectIsVisibilityHiddenIfNoScript } from '@test/helpers/no-script'
+import {
+  expectIsFlexDisplayedIfNoScript,
+  expectIsVisibilityHiddenIfNoScript,
+} from '@test/helpers/no-script'
+import { provideNoopAnimations } from '@angular/platform-browser/animations'
+import {
+  expectHiddenVisibility,
+  expectIsDisplayed,
+  expectIsNotDisplayed,
+} from '@test/helpers/visibility'
 
-describe('CollapsibleTreeNodeComponent', () => {
-  let component: CollapsibleTreeNodeComponent
-  let fixture: ComponentFixture<CollapsibleTreeNodeComponent>
+describe('CollapsibleTreeComponent', () => {
+  let component: CollapsibleTreeComponent
+  let fixture: ComponentFixture<CollapsibleTreeComponent>
 
   const DATA_CLASS_SELECTOR = By.css('.data')
 
@@ -74,6 +81,42 @@ describe('CollapsibleTreeNodeComponent', () => {
         .toEqual(DUMMY_LINE.data!.html)
     })
 
+    const LIST_ELEMENT_PREDICATE = By.css('ul')
+    it('should not render the list of elements when children list is empty', () => {
+      ;[fixture, component] = makeSut()
+      component.line = new DescriptionLine(DUMMY_LINE.data)
+
+      fixture.detectChanges()
+
+      const listElement = fixture.debugElement.query(LIST_ELEMENT_PREDICATE)
+      expect(listElement).toBeNull()
+    })
+
+    it('should render the list of elements (with assigned id and increased depth) when children list is not empty', () => {
+      const DUMMY_DEPTH = 42
+      const DUMMY_CHILDREN = [new DescriptionLine(), new DescriptionLine()]
+      ;[fixture, component] = makeSut()
+      component.line = new DescriptionLine(DUMMY_LINE.data, DUMMY_CHILDREN)
+      component.depth = DUMMY_DEPTH
+
+      fixture.detectChanges()
+
+      const listElement = fixture.debugElement.query(LIST_ELEMENT_PREDICATE)
+      const listItemElements = listElement.queryAll(By.css('li'))
+      expect(listItemElements).toHaveSize(DUMMY_CHILDREN.length)
+
+      for (const listItemElement of listItemElements) {
+        const itemElement = listItemElement.query(
+          byComponent(CollapsibleTreeComponent),
+        )
+        expect(itemElement).not.toBeNull()
+
+        expect(getReflectedAttribute(itemElement, 'depth')).toEqual(
+          (DUMMY_DEPTH + 1).toString(),
+        )
+      }
+    })
+
     describe('when collapsible', () => {
       const ALWAYS_COLLAPSIBLE: IsCollapsibleFn = () => true
       const BUTTON_PREDICATE = By.css('button')
@@ -95,32 +138,65 @@ describe('CollapsibleTreeNodeComponent', () => {
         fixture.detectChanges()
 
         const buttonElement = fixture.debugElement.query(BUTTON_PREDICATE)
-        const childrenElement = fixture.debugElement.query(
-          byComponent(CollapsibleTreeChildrenComponent),
+        expect(buttonElement.attributes[ATTRIBUTE_ARIA_CONTROLS]).toEqual(
+          component.sluggedId!,
         )
-        expect(getReflectedAttribute(childrenElement, 'id')).toEqual(
-          buttonElement.attributes[ATTRIBUTE_ARIA_CONTROLS],
-        )
+
+        const listElement = fixture.debugElement.query(LIST_ELEMENT_PREDICATE)
+        expect(listElement).not.toBeNull()
+        expect(listElement.attributes['id']).toEqual(component.sluggedId!)
       })
 
-      it('should be expanded by default when rendering on server', () => {
-        ;[fixture, component] = makeSut({
-          platformService: MOCK_SERVER_PLATFORM_SERVICE,
+      describe('when rendering on browser', () => {
+        beforeEach(() => {
+          ;[fixture, component] = makeSut({
+            platformService: MOCK_BROWSER_PLATFORM_SERVICE,
+          })
+          component.line = DUMMY_LINE
+          component.isCollapsibleFn = ALWAYS_COLLAPSIBLE
+          fixture.detectChanges()
         })
-        component.line = DUMMY_LINE
 
-        fixture.detectChanges()
-        expect(component.isExpanded).toBeTrue()
+        it('should display list', () => {
+          const listElement = fixture.debugElement.query(LIST_ELEMENT_PREDICATE)
+          expectIsDisplayed(listElement.nativeElement)
+        })
+
+        it('should be collapsed by default', () => {
+          expect(component.isExpanded).toBeFalse()
+        })
+
+        //👇 This way, if animations are deferred, the hidden visibility initial status is
+        //   already applied
+        it('should set list visibility to hidden', () => {
+          const listElement = fixture.debugElement.query(LIST_ELEMENT_PREDICATE)
+          expectHiddenVisibility(listElement.nativeElement)
+        })
       })
 
-      it('should be collapsed by default when rendering on server', () => {
-        ;[fixture, component] = makeSut({
-          platformService: MOCK_BROWSER_PLATFORM_SERVICE,
+      describe('when rendering on server', () => {
+        beforeEach(() => {
+          ;[fixture, component] = makeSut({
+            platformService: MOCK_SERVER_PLATFORM_SERVICE,
+          })
+          component.line = DUMMY_LINE
+          component.isCollapsibleFn = ALWAYS_COLLAPSIBLE
+          fixture.detectChanges()
         })
-        component.line = DUMMY_LINE
 
-        fixture.detectChanges()
-        expect(component.isExpanded).toBeFalse()
+        it('should not display list', () => {
+          const listElement = fixture.debugElement.query(LIST_ELEMENT_PREDICATE)
+          expectIsNotDisplayed(listElement.nativeElement)
+        })
+
+        it('should force list display if no JS', () => {
+          const listElement = fixture.debugElement.query(LIST_ELEMENT_PREDICATE)
+          expectIsFlexDisplayedIfNoScript(listElement)
+        })
+
+        it('should be expanded by default', () => {
+          expect(component.isExpanded).toBeTrue()
+        })
       })
 
       interface ExpandedTestCase {
@@ -128,10 +204,11 @@ describe('CollapsibleTreeNodeComponent', () => {
         readonly opposite: string
         readonly isExpanded: boolean
         readonly iconProperty: Extract<
-          keyof CollapsibleTreeNodeComponent,
+          keyof CollapsibleTreeComponent,
           'expandedIcon' | 'collapsedIcon'
         >
       }
+
       const EXPANDED_TEST_CASES = [
         {
           name: 'expanded',
@@ -148,17 +225,12 @@ describe('CollapsibleTreeNodeComponent', () => {
       ] as const satisfies ReadonlyArray<ExpandedTestCase>
       for (const testCase of EXPANDED_TEST_CASES) {
         describe(`when ${testCase.name}`, () => {
-          let mockParentComponent: CollapsibleTreeNodeComponent
-          let mockParentChildrenComponent: CollapsibleTreeChildrenComponent
+          let mockParentComponent: jasmine.SpyObj<CollapsibleTreeComponent>
 
           beforeEach(() => {
-            mockParentChildrenComponent = MockService(
-              CollapsibleTreeChildrenComponent,
-            )
-            mockParentComponent = MockService(CollapsibleTreeNodeComponent, {
-              children: mockParentChildrenComponent,
-            })
-            spyOn(mockParentChildrenComponent, 'collapseAllChildren')
+            mockParentComponent = jasmine.createSpyObj('node', [
+              'collapseAllChildren' satisfies keyof CollapsibleTreeComponent,
+            ])
             ;[fixture, component] = makeSut()
             component.line = DUMMY_LINE
             component.isCollapsibleFn = ALWAYS_COLLAPSIBLE
@@ -202,7 +274,7 @@ describe('CollapsibleTreeNodeComponent', () => {
               fixture.detectChanges()
 
               expect(
-                mockParentChildrenComponent.collapseAllChildren,
+                mockParentComponent.collapseAllChildren,
               ).toHaveBeenCalledOnceWith({ except: component })
             })
           }
@@ -214,28 +286,14 @@ describe('CollapsibleTreeNodeComponent', () => {
 
 function makeSut(
   opts: { platformService?: PlatformService } = {},
-): [
-  ComponentFixture<CollapsibleTreeNodeComponent>,
-  CollapsibleTreeNodeComponent,
-] {
-  return componentTestSetup(
-    CollapsibleTreeNodeComponent,
-    {
-      providers: [
-        MockProvider(
-          PLATFORM_SERVICE,
-          opts.platformService ?? MOCK_BROWSER_PLATFORM_SERVICE,
-        ),
-      ],
-    },
-    // Override `forwardRef` of children component
-    {
-      remove: {
-        imports: [COLLAPSIBLE_TREE_CHILDREN_COMPONENT_FORWARD_REF],
-      },
-      add: {
-        imports: [MockComponent(CollapsibleTreeChildrenComponent)],
-      },
-    },
-  )
+): [ComponentFixture<CollapsibleTreeComponent>, CollapsibleTreeComponent] {
+  return componentTestSetup(CollapsibleTreeComponent, {
+    providers: [
+      MockProvider(
+        PLATFORM_SERVICE,
+        opts.platformService ?? MOCK_BROWSER_PLATFORM_SERVICE,
+      ),
+      provideNoopAnimations(),
+    ],
+  })
 }
