+++ src/app/resume-page/profile-section/profile-description/profile-description.component.ts
@@ -2,19 +2,18 @@ import { NgClass, NgFor, NgIf, NgTemplateOutlet } from '@angular/common'
 import { Component, Inject } from '@angular/core'
 import { MaterialSymbolDirective } from '../../../common/material-symbol.directive'
 import { METADATA } from '../../../common/injection-tokens'
-import { Metadata } from '../../../metadata'
-import { CollapsibleTreeChildrenComponent } from '../../collapsible-tree/collapsible-tree-children/collapsible-tree-children.component'
+import { DescriptionLine, Metadata } from '../../../metadata'
 import {
-  CollapsibleTreeNodeComponent,
+  CollapsibleTreeComponent,
   IsCollapsibleFn,
-} from '../../collapsible-tree/collapsible-tree-node/collapsible-tree-node.component'
+} from '../../collapsible-tree/collapsible-tree.component'
 
 @Component({
   selector: 'app-profile-description',
-  template: `<app-collapsible-tree-children
-    [children]="DESCRIPTION_LINES"
+  template: `<app-collapsible-tree
+    [line]="ROOT_LINE"
     [isCollapsibleFn]="IS_COLLAPSIBLE_FN"
-  ></app-collapsible-tree-children> `,
+  ></app-collapsible-tree> `,
   standalone: true,
   imports: [
     NgIf,
@@ -22,14 +21,17 @@ import {
     MaterialSymbolDirective,
     NgFor,
     NgClass,
-    CollapsibleTreeChildrenComponent,
+    CollapsibleTreeComponent,
   ],
 })
 export class ProfileDescriptionComponent {
-  protected readonly DESCRIPTION_LINES = this.metadata.descriptionLines
+  protected readonly ROOT_LINE = new DescriptionLine(
+    undefined,
+    this.metadata.descriptionLines,
+  )
   protected readonly IS_COLLAPSIBLE_FN: IsCollapsibleFn = (
-    node: CollapsibleTreeNodeComponent,
-  ) => node.depth > 2
+    node: CollapsibleTreeComponent,
+  ) => node.depth > 1
 
   constructor(@Inject(METADATA) private readonly metadata: Metadata) {}
 }
